feat: standardize sidebar subject label formatting
