+++ solution/ui/regulations/css/scss/partials/_subjects.scss
@@ -372,6 +372,11 @@
                     border-radius: 4px;
                     padding: 4px 8px;
 
+                    &--label {
+                        display: flex;
+                        flex-direction: column;
+                    }
+
                     button {
                         width: 16px;
                         height: 16px;
