+++ solution/ui/regulations/css/scss/partials/_subjects_selector.scss
@@ -17,6 +17,21 @@
     margin-block-end: 1rem;
 }
 
+.sidebar-filters__container {
+    .sidebar__li {
+        font-size: var(--font-size-sm);
+        line-height: var(--line-height-sm);
+        padding: var(--spacer-1);
+        margin-bottom: 0;
+
+        border-bottom: 1px solid $border_color;
+
+        .subtitle {
+            display: block;
+        }
+    }
+}
+
 .subjects__select-container {
     margin-top: var(--spacer-2);
 
@@ -90,19 +105,6 @@
 
         .subjects__li {
             margin-bottom: 0.75rem;
-
-            &.sidebar__li {
-                font-size: var(--font-size-sm);
-                line-height: var(--line-height-sm);
-                padding: var(--spacer-1);
-                margin-bottom: 0;
-
-                border-bottom: 1px solid $border_color;
-
-                .subtitle {
-                    display: block;
-                }
-            }
 
             &:last-child {
                 margin-bottom: 0;
